requirements.txt
streamlit
pandas
numpy
joblib
scikit-learn
shap
langchain
langchain-google-genai
google-generativeai
matplotlib
python-dotenv
plotly
xgboost
